+++ src/Fields/Button.php
@@ -110,7 +110,7 @@ class Button extends Field_Base {
 	 */
 	public function default_sanitize_callback( mixed $value ): int {
 		// bail if value is null.
-		if ( is_null( $value ) ) {
+		if ( ! is_scalar( $value ) ) {
 			return 0;
 		}