+++ src/Fields/File.php
@@ -171,6 +171,12 @@ class File extends Field_Base {
 	 * @return int
 	 */
 	public function default_sanitize_callback( mixed $value ): int {
+		// check the value.
+		if ( ! is_scalar( $value ) ) {
+			return 0;
+		}
+
+		// return the value.
 		return absint( $value );
 	}